feat) reducer 테스트 코드 통과시키기 위해서 코드 수정
